Create Bypass-CredSSP.ps1
This is a sample file that, in this case, is used to repair data source connections on SQL Agent Jobs for cube databases.

What this script does is utilize psexec to execute a remote command.  We want the command to be dynamic.  I came up with using the "powershell.exe -EncodedCommand" as a means to pass in my dynamic code block.  The problem I encountered was that the encoded code block was too long to pass to psexec.exe, which has a character limitation.

So I write my initial base64 string to a file, and then create another base64 encoded script block that simply reads that output file, and then executes powershell.exe with the encoded string - in this manner, I can use psexec.exe with any length script block.